+++ operator/pkg/gateway-api/gateway_reconcile.go
@@ -448,6 +448,103 @@ func (r *gatewayReconciler) filterHTTPRoutesByGateway(ctx context.Context, gw *g
 	return filtered
 }
 
+func (r *gatewayReconciler) filterGRPCRoutesByGateway(ctx context.Context, gw *gatewayv1.Gateway, attachedListenerSets []gatewayv1.ListenerSet, routes []gatewayv1.GRPCRoute) []gatewayv1.GRPCRoute {
+	var filtered []gatewayv1.GRPCRoute
+	allListenerHostNames := routechecks.GetAllListenerHostNames(gw.Spec.Listeners)
+
+	for _, route := range routes {
+		if helpers.IsParentAttachable(ctx, gw, &route, route.Status.Parents, attachedListenerSets) && isAllowed(ctx, r.Client, gw, &route, r.logger) && len(computeHosts(gw, route.Spec.Hostnames, allListenerHostNames)) > 0 {
+			filtered = append(filtered, route)
+		}
+	}
+	return filtered
+}
+
+func (r *gatewayReconciler) filterHTTPRoutesByListener(ctx context.Context, gw *gatewayv1.Gateway, owner routechecks.ListenerOwner, listener *gatewayv1.Listener, routes []gatewayv1.HTTPRoute, attachedListenerSets ...gatewayv1.ListenerSet) []gatewayv1.HTTPRoute {
+	ownerNS := owner.GetNamespace()
+	var filtered []gatewayv1.HTTPRoute
+	for _, route := range routes {
+		if helpers.IsParentAttachable(ctx, gw, &route, route.Status.Parents, attachedListenerSets) &&
+			listenerisAllowed(ctx, r.Client, ownerNS, listener, &route, r.logger) &&
+			len(computeHostsForListener(listener, route.Spec.Hostnames, nil)) > 0 &&
+			parentRefMatched(owner, listener, route.GetNamespace(), route.Spec.ParentRefs) {
+			filtered = append(filtered, route)
+		}
+	}
+	return filtered
+}
+
+func (r *gatewayReconciler) filterGRPCRoutesByListener(ctx context.Context, gw *gatewayv1.Gateway, owner routechecks.ListenerOwner, listener *gatewayv1.Listener, routes []gatewayv1.GRPCRoute, attachedListenerSets ...gatewayv1.ListenerSet) []gatewayv1.GRPCRoute {
+	ownerNS := owner.GetNamespace()
+	var filtered []gatewayv1.GRPCRoute
+	for _, route := range routes {
+		if helpers.IsParentAttachable(ctx, gw, &route, route.Status.Parents, attachedListenerSets) &&
+			listenerisAllowed(ctx, r.Client, ownerNS, listener, &route, r.logger) &&
+			len(computeHostsForListener(listener, route.Spec.Hostnames, nil)) > 0 &&
+			parentRefMatched(owner, listener, route.GetNamespace(), route.Spec.ParentRefs) {
+			filtered = append(filtered, route)
+		}
+	}
+	return filtered
+}
+
+func parentRefMatched(owner routechecks.ListenerOwner, listener *gatewayv1.Listener, routeNamespace string, refs []gatewayv1.ParentReference) bool {
+	for _, ref := range refs {
+		if helpers.IsGateway(ref) {
+			// Only match if this listener belongs to a Gateway owner
+			if owner.IsListenerSet() {
+				continue
+			}
+		} else if helpers.IsListenerSet(ref) {
+			// Only match if this listener belongs to a ListenerSet owner
+			if !owner.IsListenerSet() {
+				continue
+			}
+		} else {
+			continue
+		}
+
+		if string(ref.Name) != owner.GetName() ||
+			owner.GetNamespace() != helpers.NamespaceDerefOr(ref.Namespace, routeNamespace) {
+			continue
+		}
+		if ref.SectionName == nil && ref.Port == nil {
+			return true
+		}
+		sectionNameCheck := ref.SectionName == nil || *ref.SectionName == listener.Name
+		portCheck := ref.Port == nil || *ref.Port == listener.Port
+		if sectionNameCheck && portCheck {
+			return true
+		}
+	}
+	return false
+}
+
+func (r *gatewayReconciler) filterTLSRoutesByGateway(ctx context.Context, gw *gatewayv1.Gateway, attachedListenerSets []gatewayv1.ListenerSet, routes []gatewayv1.TLSRoute) []gatewayv1.TLSRoute {
+	var filtered []gatewayv1.TLSRoute
+	for _, route := range routes {
+		if helpers.IsParentAttachable(ctx, gw, &route, route.Status.Parents, attachedListenerSets) && isAllowed(ctx, r.Client, gw, &route, r.logger) &&
+			len(computeHosts(gw, route.Spec.Hostnames, nil)) > 0 {
+			filtered = append(filtered, route)
+		}
+	}
+	return filtered
+}
+
+func (r *gatewayReconciler) filterTLSRoutesByListener(ctx context.Context, gw *gatewayv1.Gateway, owner routechecks.ListenerOwner, listener *gatewayv1.Listener, routes []gatewayv1.TLSRoute, attachedListenerSets ...gatewayv1.ListenerSet) []gatewayv1.TLSRoute {
+	ownerNS := owner.GetNamespace()
+	var filtered []gatewayv1.TLSRoute
+	for _, route := range routes {
+		if helpers.IsParentAttachable(ctx, gw, &route, route.Status.Parents, attachedListenerSets) &&
+			listenerisAllowed(ctx, r.Client, ownerNS, listener, &route, r.logger) &&
+			len(computeHostsForListener(listener, route.Spec.Hostnames, nil)) > 0 &&
+			parentRefMatched(owner, listener, route.GetNamespace(), route.Spec.ParentRefs) {
+			filtered = append(filtered, route)
+		}
+	}
+	return filtered
+}
+
 func (r *gatewayReconciler) updateListenerSetStatus(ctx context.Context, original *gatewayv1.ListenerSet, new *gatewayv1.ListenerSet) error {
 	oldStatus := original.Status.DeepCopy()
 	newStatus := new.Status.DeepCopy()
